+++ scripts/initenv.sh
@@ -9,9 +9,9 @@ if command -v lsb_release 2>&1 >/dev/null; then
         echo "You need to install python3-dev for installing the other dependencies."
         exit 1
       fi
-      if ! dpkg -l | grep build-essential 2>&1 >/dev/null
+      if ! gcc --version 2>&1 >/dev/null
       then
-        echo "You need to install build-essential for building Python dependencies."
+        echo "You need to install gcc for building Python dependencies."
         exit 1
       fi
       ;;