+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - filtermail: allow partial body length in OpenPGP payloads
   ([#570](https://github.com/chatmail/relay/pull/570))
 
+- chatmaild: allow echobot to receive unencrypted messages by default
+  ([#556](https://github.com/chatmail/relay/pull/556))
+
+
 ## 1.6.0 2025-04-11
 
 - Handle Port-25 connect errors more gracefully (common with VPNs)